requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
redis
pymongo

jinja2
python-multipart

passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]

python-dotenv
email-validator
slowapi

